Stochastic Search Inconsistency Factor Selection of brief alcohol interventions.

Stochastic Search Inconsistency Factor Selection for the evaluation of the consistency assumption for the network meta-analysis model.

These data are used as an example in Seitidis et al. (2021).

Format

A data frame with the following columns:

studyid	study id

treat1	treatment 1

treat2	treatment 2

m1	mean value of brief alcohol intervention in arm 1

m2	mean value of brief alcohol intervention in arm 2

n1	number of individuals in arm 1

n2	number of individuals in arm 2

sd1	standard deviation of brief alcohol intervention in arm 1

sd2	standard deviation of brief alcohol intervention in arm 2

TE	standardized mean difference of treat1 versus treat2

seTE	standard error of standardized mean difference

Source

Seitidis G, Nikolakopoulos S, Hennessy EA, Tanner-Smith EE, Mavridis D (2021): Network Meta-Analysis Techniques for Synthesizing Prevention Science Evidence Prevention Science, 1-10

Examples

data(Alcohol)

TE <- Alcohol$TE
seTE <- Alcohol$seTE
studlab <- Alcohol$studyid
treat1 <- Alcohol$treat2
treat2 <- Alcohol$treat1

# Stochastic Search Inconsistency Factor Selection using as reference treatment AO-CT and the
# design-by-treatment method for the specification of the Z matrix.

m <- ssifs(TE, seTE, treat1, treat2, studlab, ref = "AO-CT",
M = 1000, B = 100, M_pilot = 1000, B_pilot = 100)